FireScreen: Refactor to separate the fire simulation logic from screen-specific UI composition.
Modifies:
- TUI.vcxproj
- Screens/FireScreen.h/.cpp

Adds:
- Screens/Content/FireEffect.h/.cpp

- Turns the fire sim logic into a reusable fire effect
- Now renders into a caller-provided viewport (Rect)

- FireScreen
  - Owns all UI/chrome:
    - outer/inner frames
    - title/footer bars
    - banner/logo
    - warning messages
  - Defines content area and passes it to FireEffect::draw()

Architectural impact:

- Establishes clear separation between:
  - "Effect" (what is rendered)
  - "Screen/UI" (where and how it is presented)

- Enables FireEffect to be reused in:
  - windows
  - popups
  - future widgets
  - layered/composited UI
